+++ pkg/dev/kubedev/components.go
@@ -7,7 +7,8 @@ import (
 	"path/filepath"
 	"reflect"
 	"strings"
-	"time"
+
+	"golang.org/x/sync/errgroup"
 
 	devfilev1 "github.com/devfile/api/v2/pkg/apis/workspaces/v1alpha2"
 	"github.com/devfile/library/v2/pkg/devfile/generator"
@@ -15,7 +16,6 @@ import (
 	parsercommon "github.com/devfile/library/v2/pkg/devfile/parser/data/v2/common"
 	devfilefs "github.com/devfile/library/v2/pkg/testingutil/filesystem"
 	dfutil "github.com/devfile/library/v2/pkg/util"
-	"golang.org/x/sync/errgroup"
 
 	"github.com/redhat-developer/odo/pkg/component"
 	"github.com/redhat-developer/odo/pkg/dev/common"
@@ -27,10 +27,8 @@ import (
 	"github.com/redhat-developer/odo/pkg/libdevfile"
 	"github.com/redhat-developer/odo/pkg/log"
 	odocontext "github.com/redhat-developer/odo/pkg/odo/context"
-	"github.com/redhat-developer/odo/pkg/port"
 	"github.com/redhat-developer/odo/pkg/service"
 	storagepkg "github.com/redhat-developer/odo/pkg/storage"
-	"github.com/redhat-developer/odo/pkg/sync"
 	"github.com/redhat-developer/odo/pkg/testingutil/filesystem"
 	"github.com/redhat-developer/odo/pkg/util"
 	"github.com/redhat-developer/odo/pkg/watch"
@@ -44,27 +42,23 @@ import (
 	"k8s.io/utils/pointer"
 )
 
-// Push updates the component if a matching component exists or creates one if it doesn't exist
-// Once the component has started, it will sync the source code to it.
-// The componentStatus will be modified to reflect the status of the component when the function returns
-func (o *DevClient) Push(ctx context.Context, parameters common.PushParameters, componentStatus *watch.ComponentStatus) (err error) {
-
+// createComponents creates the components into the cluster
+// returns true if the pod is created
+func (o *DevClient) createComponents(ctx context.Context, parameters common.PushParameters, componentStatus *watch.ComponentStatus) (bool, error) {
 	var (
 		appName       = odocontext.GetApplication(ctx)
 		componentName = odocontext.GetComponentName(ctx)
-		devfilePath   = odocontext.GetDevfilePath(ctx)
-		path          = filepath.Dir(devfilePath)
 	)
 
 	// preliminary checks
-	err = dfutil.ValidateK8sResourceName("component name", componentName)
+	err := dfutil.ValidateK8sResourceName("component name", componentName)
 	if err != nil {
-		return err
+		return false, err
 	}
 
 	err = dfutil.ValidateK8sResourceName("component namespace", o.kubernetesClient.GetCurrentNamespace())
 	if err != nil {
-		return err
+		return false, err
 	}
 
 	if componentStatus.State == watch.StateSyncOutdated {
@@ -75,12 +69,13 @@ func (o *DevClient) Push(ctx context.Context, parameters common.PushParameters, 
 	klog.V(4).Infof("component state: %q\n", componentStatus.State)
 	err = o.buildPushAutoImageComponents(ctx, o.filesystem, parameters.Devfile, componentStatus)
 	if err != nil {
-		return err
+		return false, err
 	}
 
-	deployment, deploymentExists, err := o.getComponentDeployment(ctx)
+	var deployment *appsv1.Deployment
+	deployment, o.deploymentExists, err = o.getComponentDeployment(ctx)
 	if err != nil {
-		return err
+		return false, err
 	}
 
 	if componentStatus.State != watch.StateWaitDeployment && componentStatus.State != watch.StateReady {
@@ -92,13 +87,13 @@ func (o *DevClient) Push(ctx context.Context, parameters common.PushParameters, 
 	labels := odolabels.GetLabels(componentName, appName, runtime, odolabels.ComponentDevMode, false)
 
 	var updated bool
-	deployment, updated, err = o.createOrUpdateComponent(ctx, parameters, deploymentExists, libdevfile.DevfileCommands{
+	deployment, updated, err = o.createOrUpdateComponent(ctx, parameters, o.deploymentExists, libdevfile.DevfileCommands{
 		BuildCmd: parameters.DevfileBuildCmd,
 		RunCmd:   parameters.DevfileRunCmd,
 		DebugCmd: parameters.DevfileDebugCmd,
 	}, deployment)
 	if err != nil {
-		return fmt.Errorf("unable to create or update component: %w", err)
+		return false, fmt.Errorf("unable to create or update component: %w", err)
 	}
 	ownerReference := generator.GetOwnerReference(deployment)
 
@@ -107,12 +102,12 @@ func (o *DevClient) Push(ctx context.Context, parameters common.PushParameters, 
 
 	objectsToRemove, serviceBindingSecretsToRemove, err := o.getRemoteResourcesNotPresentInDevfile(ctx, parameters, selector)
 	if err != nil {
-		return fmt.Errorf("unable to determine resources to delete: %w", err)
+		return false, fmt.Errorf("unable to determine resources to delete: %w", err)
 	}
 
 	err = o.deleteRemoteResources(objectsToRemove)
 	if err != nil {
-		return fmt.Errorf("unable to delete remote resources: %w", err)
+		return false, fmt.Errorf("unable to delete remote resources: %w", err)
 	}
 
 	// this is mainly useful when the Service Binding Operator is not installed;
@@ -120,212 +115,56 @@ func (o *DevClient) Push(ctx context.Context, parameters common.PushParameters, 
 	if len(serviceBindingSecretsToRemove) != 0 {
 		err = o.deleteServiceBindingSecrets(serviceBindingSecretsToRemove, deployment)
 		if err != nil {
-			return fmt.Errorf("unable to delete service binding secrets: %w", err)
+			return false, fmt.Errorf("unable to delete service binding secrets: %w", err)
 		}
 	}
 
 	// Create all the K8s components defined in the devfile
 	_, err = o.pushDevfileKubernetesComponents(ctx, parameters, labels, odolabels.ComponentDevMode, ownerReference)
 	if err != nil {
-		return err
+		return false, err
 	}
 
 	err = o.updatePVCsOwnerReferences(ctx, ownerReference)
 	if err != nil {
-		return err
+		return false, err
 	}
 
 	if updated {
 		klog.V(4).Infof("Deployment has been updated to generation %d. Waiting new event...\n", deployment.GetGeneration())
 		componentStatus.State = watch.StateWaitDeployment
-		return nil
+		return false, nil
 	}
 
 	numberReplicas := deployment.Status.ReadyReplicas
 	if numberReplicas != 1 {
 		klog.V(4).Infof("Deployment has %d ready replicas. Waiting new event...\n", numberReplicas)
 		componentStatus.State = watch.StateWaitDeployment
-		return nil
+		return false, nil
 	}
 
 	injected, err := o.bindingClient.CheckServiceBindingsInjectionDone(componentName, appName)
 	if err != nil {
-		return err
+		return false, err
 	}
 
 	if !injected {
 		klog.V(4).Infof("Waiting for all service bindings to be injected...\n")
-		return errors.New("some servicebindings are not injected")
+		return false, errors.New("some servicebindings are not injected")
 	}
 
 	// Check if endpoints changed in Devfile
 	portsToForward, err := libdevfile.GetDevfileContainerEndpointMapping(parameters.Devfile, parameters.Debug)
 	if err != nil {
-		return err
+		return false, err
 	}
-	portsChanged := !reflect.DeepEqual(portsToForward, o.portForwardClient.GetForwardedPorts())
+	o.portsChanged = !reflect.DeepEqual(portsToForward, o.portForwardClient.GetForwardedPorts())
 
-	if componentStatus.State == watch.StateReady && !portsChanged {
+	if componentStatus.State == watch.StateReady && !o.portsChanged {
 		// If the deployment is already in Ready State, no need to continue
-		return nil
+		return false, nil
 	}
-
-	// Now the Deployment has a Ready replica, we can get the Pod to work inside it
-	pod, err := o.kubernetesClient.GetPodUsingComponentName(componentName)
-	if err != nil {
-		return fmt.Errorf("unable to get pod for component %s: %w", componentName, err)
-	}
-
-	// Find at least one pod with the source volume mounted, error out if none can be found
-	containerName, syncFolder, err := common.GetFirstContainerWithSourceVolume(pod.Spec.Containers)
-	if err != nil {
-		return fmt.Errorf("error while retrieving container from pod %s with a mounted project volume: %w", pod.GetName(), err)
-	}
-
-	s := log.Spinner("Syncing files into the container")
-	defer s.End(false)
-
-	// Get commands
-	pushDevfileCommands, err := o.getPushDevfileCommands(parameters)
-	if err != nil {
-		return fmt.Errorf("failed to validate devfile build and run commands: %w", err)
-	}
-
-	podChanged := componentStatus.State == watch.StateWaitDeployment
-
-	// Get a sync adapter. Check if project files have changed and sync accordingly
-	compInfo := sync.ComponentInfo{
-		ComponentName: componentName,
-		ContainerName: containerName,
-		PodName:       pod.GetName(),
-		SyncFolder:    syncFolder,
-	}
-
-	cmdKind := devfilev1.RunCommandGroupKind
-	cmdName := parameters.DevfileRunCmd
-	if parameters.Debug {
-		cmdKind = devfilev1.DebugCommandGroupKind
-		cmdName = parameters.DevfileDebugCmd
-	}
-
-	syncParams := sync.SyncParameters{
-		Path:                     path,
-		WatchFiles:               parameters.WatchFiles,
-		WatchDeletedFiles:        parameters.WatchDeletedFiles,
-		IgnoredFiles:             parameters.IgnoredFiles,
-		DevfileScanIndexForWatch: parameters.DevfileScanIndexForWatch,
-
-		CompInfo:  compInfo,
-		ForcePush: !deploymentExists || podChanged,
-		Files:     common.GetSyncFilesFromAttributes(pushDevfileCommands[cmdKind]),
-	}
-
-	execRequired, err := o.syncClient.SyncFiles(ctx, syncParams)
-	if err != nil {
-		componentStatus.State = watch.StateReady
-		return fmt.Errorf("failed to sync to component with name %s: %w", componentName, err)
-	}
-	s.End(true)
-
-	// PostStart events from the devfile will only be executed when the component
-	// didn't previously exist
-	if !componentStatus.PostStartEventsDone && libdevfile.HasPostStartEvents(parameters.Devfile) {
-		err = libdevfile.ExecPostStartEvents(ctx, parameters.Devfile, component.NewExecHandler(o.kubernetesClient, o.execClient, appName, componentName, pod.Name, "Executing post-start command in container", parameters.Show))
-		if err != nil {
-			return err
-		}
-	}
-	componentStatus.PostStartEventsDone = true
-
-	cmd, err := libdevfile.ValidateAndGetCommand(parameters.Devfile, cmdName, cmdKind)
-	if err != nil {
-		return err
-	}
-
-	commandType, err := parsercommon.GetCommandType(cmd)
-	if err != nil {
-		return err
-	}
-	var running bool
-	var isComposite bool
-	cmdHandler := runHandler{
-		fs:            o.filesystem,
-		execClient:    o.execClient,
-		kubeClient:    o.kubernetesClient,
-		appName:       appName,
-		componentName: componentName,
-		devfile:       parameters.Devfile,
-		path:          path,
-		podName:       pod.GetName(),
-		ctx:           ctx,
-	}
-
-	if commandType == devfilev1.ExecCommandType {
-		running, err = cmdHandler.IsRemoteProcessForCommandRunning(ctx, cmd, pod.Name)
-		if err != nil {
-			return err
-		}
-	} else if commandType == devfilev1.CompositeCommandType {
-		// this handler will run each command in this composite command individually,
-		// and will determine whether each command is running or not.
-		isComposite = true
-	} else {
-		return fmt.Errorf("unsupported type %q for Devfile command %s, only exec and composite are handled",
-			commandType, cmd.Id)
-	}
-
-	cmdHandler.componentExists = running || isComposite
-
-	klog.V(4).Infof("running=%v, execRequired=%v",
-		running, execRequired)
-
-	if isComposite || !running || execRequired {
-		// Invoke the build command once (before calling libdevfile.ExecuteCommandByNameAndKind), as, if cmd is a composite command,
-		// the handler we pass will be called for each command in that composite command.
-		doExecuteBuildCommand := func() error {
-			execHandler := component.NewExecHandler(o.kubernetesClient, o.execClient, appName, componentName, pod.Name,
-				"Building your application in container", parameters.Show)
-			return libdevfile.Build(ctx, parameters.Devfile, parameters.DevfileBuildCmd, execHandler)
-		}
-		if running {
-			if cmd.Exec == nil || !util.SafeGetBool(cmd.Exec.HotReloadCapable) {
-				if err = doExecuteBuildCommand(); err != nil {
-					return err
-				}
-			}
-		} else {
-			if err = doExecuteBuildCommand(); err != nil {
-				return err
-			}
-		}
-		err = libdevfile.ExecuteCommandByNameAndKind(ctx, parameters.Devfile, cmdName, cmdKind, &cmdHandler, false)
-		if err != nil {
-			return err
-		}
-	}
-
-	if podChanged || portsChanged {
-		o.portForwardClient.StopPortForwarding(ctx, componentName)
-	}
-
-	// Check that the application is actually listening on the ports declared in the Devfile, so we are sure that port-forwarding will work
-	appReadySpinner := log.Spinner("Waiting for the application to be ready")
-	err = o.checkAppPorts(ctx, pod.Name, portsToForward)
-	appReadySpinner.End(err == nil)
-	if err != nil {
-		log.Warningf("Port forwarding might not work correctly: %v", err)
-		log.Warning("Running `odo logs --follow` might help in identifying the problem.")
-		fmt.Fprintln(log.GetStdout())
-	}
-
-	err = o.portForwardClient.StartPortForwarding(ctx, parameters.Devfile, componentName, parameters.Debug, parameters.RandomPorts, log.GetStdout(), parameters.ErrOut, parameters.CustomForwardedPorts)
-	if err != nil {
-		return common.NewErrPortForward(err)
-	}
-	componentStatus.EndpointsForwarded = o.portForwardClient.GetForwardedPorts()
-
-	componentStatus.State = watch.StateReady
-	return nil
+	return true, nil
 }
 
 func (o *DevClient) buildPushAutoImageComponents(ctx context.Context, fs filesystem.Filesystem, devfileObj parser.DevfileObj, compStatus *watch.ComponentStatus) error {
@@ -774,33 +613,6 @@ func (o *DevClient) updatePVCsOwnerReferences(ctx context.Context, ownerReferenc
 		}
 	}
 	return nil
-}
-
-func (o *DevClient) getPushDevfileCommands(parameters common.PushParameters) (map[devfilev1.CommandGroupKind]devfilev1.Command, error) {
-	pushDevfileCommands, err := libdevfile.ValidateAndGetPushCommands(parameters.Devfile, parameters.DevfileBuildCmd, parameters.DevfileRunCmd)
-	if err != nil {
-		return nil, fmt.Errorf("failed to validate devfile build and run commands: %w", err)
-	}
-
-	if parameters.Debug {
-		pushDevfileDebugCommands, e := libdevfile.ValidateAndGetCommand(parameters.Devfile, parameters.DevfileDebugCmd, devfilev1.DebugCommandGroupKind)
-		if e != nil {
-			return nil, fmt.Errorf("debug command is not valid: %w", e)
-		}
-		pushDevfileCommands[devfilev1.DebugCommandGroupKind] = pushDevfileDebugCommands
-	}
-
-	return pushDevfileCommands, nil
-}
-
-func (o *DevClient) checkAppPorts(ctx context.Context, podName string, portsToFwd map[string][]devfilev1.Endpoint) error {
-	containerPortsMapping := make(map[string][]int)
-	for c, ports := range portsToFwd {
-		for _, p := range ports {
-			containerPortsMapping[c] = append(containerPortsMapping[c], p.TargetPort)
-		}
-	}
-	return port.CheckAppPortsListening(ctx, o.execClient, podName, containerPortsMapping, 1*time.Minute)
 }
 
 // generateDeploymentObjectMeta generates a ObjectMeta object for the given deployment's name, labels and annotations
